requirements.txt
flask==3.0.3
werkzeug==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-limiter==3.8.0
